+++ sdk/formrecognizer/Azure.AI.FormRecognizer/src/FieldBoundingBox.cs
@@ -16,11 +16,11 @@ namespace Azure.AI.FormRecognizer.Models
     /// relative to the text orientation. Units are in pixels for images and inches for PDF. The
     /// <see cref="LengthUnit"/> type of a recognized page can be found at <see cref="FormPage.Unit"/>.
     /// </summary>
-    public readonly struct BoundingBox
+    public readonly struct FieldBoundingBox
     {
         private readonly PointF[] _points;
 
-        internal BoundingBox(IReadOnlyList<float> boundingBox)
+        internal FieldBoundingBox(IReadOnlyList<float> boundingBox)
         {
             if (boundingBox.Count == 0)
             {
@@ -38,10 +38,10 @@ namespace Azure.AI.FormRecognizer.Models
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="BoundingBox"/> structure.
+        /// Initializes a new instance of the <see cref="FieldBoundingBox"/> structure.
         /// </summary>
-        /// <param name="points">The sequence of points defining this <see cref="BoundingBox"/>.</param>
-        internal BoundingBox(IReadOnlyList<PointF> points)
+        /// <param name="points">The sequence of points defining this <see cref="FieldBoundingBox"/>.</param>
+        internal FieldBoundingBox(IReadOnlyList<PointF> points)
         {
             _points = points?.ToArray();
         }
@@ -51,7 +51,7 @@ namespace Azure.AI.FormRecognizer.Models
         internal PointF[] Points => _points ?? Array.Empty<PointF>();
 
         /// <summary>
-        /// Gets one of the points that set the limits of this <see cref="BoundingBox"/>.
+        /// Gets one of the points that set the limits of this <see cref="FieldBoundingBox"/>.
         /// Coordinates are specified relative to the top-left of the original image, and points
         /// are ordered clockwise from the top-left corner relative to the text orientation.
         /// </summary>
@@ -60,7 +60,7 @@ namespace Azure.AI.FormRecognizer.Models
         public PointF this[int index] => Points[index];
 
         /// <summary>
-        /// Returns string representation for <see cref="BoundingBox"/>.
+        /// Returns string representation for <see cref="FieldBoundingBox"/>.
         /// </summary>
         public override string ToString() => string.Join(",", Points.Select(p => p.ToString()).ToArray());
     }