+++ sdk/formrecognizer/Azure.AI.FormRecognizer/tests/Models/FieldBoundingBoxTests.cs
@@ -9,35 +9,35 @@ using NUnit.Framework;
 namespace Azure.AI.FormRecognizer.Tests
 {
     /// <summary>
-    /// The suite of tests for the <see cref="BoundingBox"/> struct.
+    /// The suite of tests for the <see cref="FieldBoundingBox"/> struct.
     /// </summary>
-    public class BoundingBoxTests
+    public class FieldBoundingBoxTests
     {
         [Test]
-        public void IndexerThrowsWhenBoundingBoxIsDefault()
+        public void IndexerThrowsWhenFieldBoundingBoxIsDefault()
         {
-            BoundingBox boundingBox = default;
+            FieldBoundingBox boundingBox = default;
             Assert.Throws<IndexOutOfRangeException>(() => { var _ = boundingBox[0]; });
         }
 
         [Test]
-        public void IndexerThrowsWhenBoundingBoxIsEmpty()
+        public void IndexerThrowsWhenFieldBoundingBoxIsEmpty()
         {
-            BoundingBox boundingBox = new BoundingBox(new List<float>());
+            FieldBoundingBox boundingBox = new FieldBoundingBox(new List<float>());
             Assert.Throws<IndexOutOfRangeException>(() => { var _ = boundingBox[0]; });
         }
 
         [Test]
-        public void ToStringDoesNotThrowWhenBoundingBoxIsDefault()
+        public void ToStringDoesNotThrowWhenFieldBoundingBoxIsDefault()
         {
-            BoundingBox boundingBox = default;
+            FieldBoundingBox boundingBox = default;
             Assert.DoesNotThrow(() => boundingBox.ToString());
         }
 
         [Test]
-        public void ToStringDoesNotThrowWhenBoundingBoxIsEmpty()
+        public void ToStringDoesNotThrowWhenFieldBoundingBoxIsEmpty()
         {
-            BoundingBox boundingBox = new BoundingBox(new List<float>());
+            FieldBoundingBox boundingBox = new FieldBoundingBox(new List<float>());
             Assert.DoesNotThrow(() => boundingBox.ToString());
         }
     }